client/requirements.txt
# Framework principal para desenvolvimento do jogo 2D
pygame==2.5.2

# Rede e comunicação com servidor LAN
python-socketio==5.10.0
eventlet==0.33.3

# Banco de dados e ORM
sqlite3==3.37.0  # Geralmente já incluído no Python
sqlalchemy==2.0.30
alembic==1.13.1

# Serialização de dados para comunicação em rede
python-engineio==4.9.0

# Processamento de dados para IA
numpy==1.26.4

# Configurações e arquivos YAML
pyyaml==6.0.1

# Utilitários e ferramentas auxiliares
pillow==10.3.0  # Processamento de imagens
python-dateutil==2.9.0.post0  # Manipulação de datas